Update servicenow-tickets-connector.md
As written, the user will experience and error when attempting to authenticate with ServiceNow. The error is "Can't authenticate with the data source. verify that the credentials associated with this data source are correct and that the data source path is valid. The credentials do not have access to the following tables: sys_dp_object."

By adding row and column permission to the sys_db_object table, I no longer received the error. Moved 'sys_db_object' from optional to required tables list.